+++ aiuser/response/response.py
@@ -10,7 +10,7 @@ from redbot.core import Config, commands
 from aiuser.config.constants import REGEX_RUN_TIMEOUT
 from aiuser.context.messages import MessagesThread
 from aiuser.context.setup import create_messages_thread
-from aiuser.response.chat.llm_pipeline import LLMPipeline
+from aiuser.response.llm_pipeline import LLMPipeline
 from aiuser.types.abc import MixinMeta
 from aiuser.utils.utilities import to_thread
